Use traits for the public api
